Set Quality of Service (QoS) for GCD queue in OS X 10.10+
If we don't set this the queue is treated as QOS_CLASS_UNSPECIFIED. Going with
QOS_CLASS_USER_INTERACTIVE for now, most aggressive. As for the check, using the
new API availability feature in Swift 2.0. This is a runtime check, but we only
do it once. Prefer this over a compiler macro, as we get a "universal" binary.