Add IOReturn.swift from SystemKit
SMCKit no longer lists all IOReturn codes, so were using IOReturn.swift from
SystemKit instead. However, we get a namespace conflict this way, so updating
SystemKit submodule as a quick workaround.

SystemKit needs refactoring and should be wrapping the IOReturn error codes. So
this is fine since it's all going to be temporary.